added checks

+++ makeUSB.sh
@@ -6,19 +6,39 @@ if [ "$(id -u)" != "0" ]; then
    exit 1
 fi
 
+#get tools
+if [ ! -f "/usr/lib/syslinux/mbr/mbr.bin" ]
+then
+  echo "Installing needed packages..."
+  apt-get install live-boot live-boot-initramfs-tools extlinux
+  update-initramfs -u
+fi
+
+echo "Where is the drive mounted?"
+echo "Example /mnt"
+read TARGET
+
+if [ TARGET = "" ]
+then
+    echo "A Mounted Location is Required"
+    exit 1
+fi
+
 #variables
 url="http://download.grml.org/grml32-small_2014.11.iso"
 ISO="/tmp/grml.iso"
-TARGET="/mnt"
 mnt=/tmp/mnt
 
 echo "Which Device do you want to install to?"
 echo "Example: /dev/sdb"
 read dev
 
-#get tools
-apt-get install live-boot live-boot-initramfs-tools extlinux
-update-initramfs -u
+if [ dev = "" ]
+then
+  echo "A Device is needed"
+  exit 1
+fi
+
 
 #get ISO
 wget -c "$url" -O "$ISO"
